CTCTRADERS-1058: wip extracting listnames from JsObject
