+++ app/services/ingestion/SchemaValidationService.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package services
+package services.ingestion
 
 import java.io.ByteArrayInputStream
