+++ test/controllers/ReferenceDataControllerSpec.scala
@@ -32,9 +32,10 @@ import play.api.libs.json.Json
 import play.api.mvc.AnyContentAsRaw
 import play.api.test.FakeRequest
 import play.api.test.Helpers._
-import services.ReferenceDataService.DataProcessingResult._
-import services.ReferenceDataService
-import services.SchemaValidationService
+import services.ingestion.ReferenceDataService.DataProcessingResult.DataProcessingFailed
+import services.ingestion.ReferenceDataService.DataProcessingResult.DataProcessingSuccessful
+import services.ingestion.ReferenceDataService
+import services.ingestion.SchemaValidationService
 
 import scala.concurrent.Future
 
@@ -53,7 +54,7 @@ class ReferenceDataControllerSpec extends SpecBase with GuiceOneAppPerSuite with
 
     "returns ACCEPTED when the data has been decompressed, validated and processed" in {
       when(mockReferenceDataService.validateAndDecompress(any(), any())).thenReturn(Right(testJson))
-      when(mockReferenceDataService.insert(any(), any())).thenReturn(Future.successful(DataProcessingSuccessful))
+      when(mockReferenceDataService.insert(any())).thenReturn(Future.successful(DataProcessingSuccessful))
 
       val result = route(app, fakeRequest).value
 
@@ -62,7 +63,7 @@ class ReferenceDataControllerSpec extends SpecBase with GuiceOneAppPerSuite with
 
     "returns Bad Request when a decompression or validation error occurs" in {
       when(mockReferenceDataService.validateAndDecompress(any(), any())).thenReturn(Left(OtherError("error")))
-      when(mockReferenceDataService.insert(any(), any())).thenReturn(Future.successful(DataProcessingSuccessful))
+      when(mockReferenceDataService.insert(any())).thenReturn(Future.successful(DataProcessingSuccessful))
 
       val result = route(app, fakeRequest).value
 
@@ -71,7 +72,7 @@ class ReferenceDataControllerSpec extends SpecBase with GuiceOneAppPerSuite with
 
     "returns with an Internal Server Error when the has been validated but data was not processed successfully" in {
       when(mockReferenceDataService.validateAndDecompress(any(), any())).thenReturn(Right(testJson))
-      when(mockReferenceDataService.insert(any(), any())).thenReturn(Future.successful(DataProcessingFailed))
+      when(mockReferenceDataService.insert(any())).thenReturn(Future.successful(DataProcessingFailed))
 
       val result = route(app, fakeRequest).value
 
@@ -88,7 +89,7 @@ class ReferenceDataControllerSpec extends SpecBase with GuiceOneAppPerSuite with
 
     "returns ACCEPTED when the data has been decompressed, validated and processed" in {
       when(mockReferenceDataService.validateAndDecompress(any(), any())).thenReturn(Right(testJson))
-      when(mockReferenceDataService.insert(any(), any())).thenReturn(Future.successful(DataProcessingSuccessful))
+      when(mockReferenceDataService.insert(any())).thenReturn(Future.successful(DataProcessingSuccessful))
 
       val result = route(app, fakeRequest).value
 
@@ -97,7 +98,7 @@ class ReferenceDataControllerSpec extends SpecBase with GuiceOneAppPerSuite with
 
     "returns Bad Request when a decompression or validation error occurs" in {
       when(mockReferenceDataService.validateAndDecompress(any(), any())).thenReturn(Left(OtherError("error")))
-      when(mockReferenceDataService.insert(any(), any())).thenReturn(Future.successful(DataProcessingSuccessful))
+      when(mockReferenceDataService.insert(any())).thenReturn(Future.successful(DataProcessingSuccessful))
 
       val result = route(app, fakeRequest).value
 
@@ -106,7 +107,7 @@ class ReferenceDataControllerSpec extends SpecBase with GuiceOneAppPerSuite with
 
     "returns with an Internal Server Error when the has been validated but data was not processed successfully" in {
       when(mockReferenceDataService.validateAndDecompress(any(), any())).thenReturn(Right(testJson))
-      when(mockReferenceDataService.insert(any(), any())).thenReturn(Future.successful(DataProcessingFailed))
+      when(mockReferenceDataService.insert(any())).thenReturn(Future.successful(DataProcessingFailed))
 
       val result = route(app, fakeRequest).value
