+++ test/resources/test.application.conf
@@ -14,7 +14,7 @@
 
 include "application.conf"
 
-play.modules.enabled += "services.TestModule"
+play.modules.enabled += "services.ingestion.TestModule"
 
 mongodb {
   uri = "mongodb://localhost:27017/customs-reference-data-test"
