+++ test/services/ingestion/GZipServiceSpec.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package services
+package services.ingestion
 
 import base.SpecBase
 import models.OtherError
